+++ JotTouchSDK.framework/Versions/A/Headers/JotSettingsViewController.h
@@ -9,6 +9,15 @@
 #import <UIKit/UIKit.h>
 
 @interface JotSettingsViewController : UINavigationController
+
+/*! Initialization of settings view controller with or without adonit jot family toggle.
+ * \param showSwitch A boolean that determines whether to show jot family toggle.
+ */
 -(id)initWithOnOffSwitch:(BOOL)showSwitch;
+
+/*! Initialization of settings view controller with or without adonit jot family toggle and palm rejection toggle.
+ * \param showSwitch A boolean that determines whether to show jot family toggle
+ * \param showPalmRejection A boolean that determines whether to show palm rejection toggle
+ */
 -(id)initWithOnOffSwitch:(BOOL)showSwitch andShowPalmRejection:(BOOL)showPalmRejection;
 @end