+++ JotTouchSDK.framework/Versions/A/Headers/JotStylusConnectionDelegate.h
@@ -2,7 +2,7 @@
 //  StylusConnectionDelegate.h
 //  PalmRejectionExampleApp
 //
-//  Created by UM on 9/14/12.
+//  Created on 9/14/12.
 //  Copyright (c) 2012 Adonit. All rights reserved.
 //
 
@@ -12,8 +12,24 @@
 @protocol JotStylusConnection;
 
 @protocol JotStylusConnectionDelegate<NSObject>
+
+/*! Sent to update the status of stylus to pairing.
+ */
+-(void)jotStylusPairing;
+
+/*! Sent to update the status of stylus to connected.
+ */
 -(void)jotStylusConnected;
+
+/*! Sent to update the status of stylus to disconnected.
+ */
 -(void)jotStylusDisconnected;
+
+/*! Sent to update the level of battery remaining.
+ * \param batteryLevel Positive integer specifying the remaining battery of connected device
+ */
 -(void)jotStylusBatteryUpdate:(uint)batteryLevel;
+
+
 @end