+++ JotTouchSDK.framework/Versions/A/Headers/JotStylusManager.h
@@ -2,7 +2,7 @@
 //  JotStylusManager.h
 //  PalmRejectionExampleApp
 //
-//  Created by me on 8/20/12.
+//  Created on 8/20/12.
 //  Copyright (c) 2012 Adonit. All rights reserved.
 //
 
@@ -10,38 +10,15 @@
 #import "JotStylusConnectionDelegate.h"
 #import "JotPalmRejectionDelegate.h"
 #import "JotStylusStateDelegate.h"
+#import "JotPreferredStylus.h"
 #import <CoreGraphics/CoreGraphics.h>
 #import <UIKit/UIKit.h>
 #import "JotShortcut.h"
 #import "JotSettingsViewController.h"
 
-#define JOT_MIN_PRESSURE 0
-#define JOT_MAX_PRESSURE 2047
-
-#define JotStylusManagerDidChangeConnectionStatus @"jotStylusManagerDidChangeConnectionStatus"
-#define JotStylusManagerDidChangeBatteryLevel @"jotStylusManagerDidChangeBatteryLevel"
-
-typedef enum {
-    JotPalmRejectionLeftHanded,
-    JotPalmRejectionRightHanded
-} JotPalmRejectionOrientation;
-
-typedef enum {
-    JotConnectionStatusOff,
-    JotConnectionStatusScanning,
-    JotConnectionStatusPairing,
-    JotConnectionStatusConnected,
-    JotConnectionStatusDisconnected
-} JotConnectionStatus;
-
-typedef enum {
-    JotPreferredStylusBT21,
-    JotPreferredStylusBT40,
-    JotNoPreferredStylus,
-} JotPreferredStylusType;
-
 
 @interface JotStylusManager : NSObject<JotStylusStateDelegate,JotStylusConnectionDelegate>
+
 + (JotStylusManager*)sharedInstance;
 
 -(void)touchesBegan:(NSSet *)touches;
@@ -49,32 +26,120 @@ typedef enum {
 -(void)touchesEnded:(NSSet *)touches;
 -(void)touchesCancelled:(NSSet *)touches;
 
-
+/*! Adds a shortcut option that is accessibile and can be specified by the user.
+ * \param shortcut A shortcut to be added to the settings interface
+ */
 -(void)addShortcutOption:(JotShortcut *)shortcut;
+
+/*! Sets the default option state for the first shortcut that will used when initially loading the interface.
+ * \param shortcut The default option of the second stylus button shortcut to be added to the settings interface
+ */
 -(void)addShortcutOptionButton1Default:(JotShortcut *)shortcut;
+
+/*! Sets the default option state for the second shortcut that will used when initially loading the interface.
+ * \param shortcut The default option of the second stylus button shortcut to be added to the settings interface
+ */
 -(void)addShortcutOptionButton2Default:(JotShortcut *)shortcut;
 
+/*! Obtains pressure data of the stylus currently connected.
+ * \returns If connected, returns positive integer value of connected pressure data. If not connected, returns the unconnected pressure data.
+ */
 -(uint)getPressure;
+
+/*! Determines if a stylus is currently connected.
+ * \returns A boolean specifying if a stylus is connected
+ */
 -(BOOL)isStylusConnected;
+
+/*! Number of stylus connected to BT, including those still pairing
+ * \returns A uint with the total count of stylus connected and/or pairing
+ */
+-(uint)totalNumberOfStylusesConnected;
+
+
+/*! Removes the current connected stylus and stop receiving data from it. 
+ */
 -(void)forgetAndTurnOffStylus;
 
+/*! Sets a view to receive touch events from Jot styluses.
+ * \param view A view that will be supplied touch events from Jot styluses
+ */
 -(void)registerView:(UIView*)view;
+
+/*! Removes view from receiving touch events from Jot styluses.
+ * \param view A view that will no longer receiver touch events from Jot styluses
+ */
 -(void)unregisterView:(UIView*)view;
 
+/*! Links to Safari and appropriate help site for the stylus currently connected.
+ */
 -(void)launchHelp;
 
+-(void)setOptionValue:(id)value forKey:(NSString *)key;
+
+#pragma mark - properties
+
+/*! Delays used to tune re-enabling gestures when a stylus is lifted from the screen.
+ */
+@property (readwrite) CGFloat palmDetectionTouchDelay;
+@property (nonatomic) CGFloat palmDetectionTouchDelayNoPressure;
 @property (readwrite) uint unconnectedPressure;
 
+/*! Array of JotShortcuts utilized in the settings interface.
+ */
 @property (readonly) NSArray *shortcuts;
+
+/*! The current button 1 shortcut of the preferred stylus.
+ */
 @property (readwrite,assign) JotShortcut *button1Shortcut;
+
+/*! The current button 2 shortcut of the preferred stylus.
+ */
 @property (readwrite,assign) JotShortcut *button2Shortcut;
 
+/*! Palm rejection delegate capturing touch events for palm rejection.
+ */
 @property (readwrite,assign) id<JotPalmRejectionDelegate> palmRejectorDelegate;
+
 @property (readwrite) BOOL enabled;
+
+/*! A string representation of the current version of the SDK being used.
+ */
 @property (readonly) NSString *SDKVersion;
+
+/*! A string representation of the current build number of the SDK being used.
+ */
+@property (readonly) NSString *SDKBuildVersion;
+
+/*! A boolean specifying whether palm rejection is on.
+ */
 @property (readwrite) BOOL rejectMode;
+
+/*! A positive integer specifying the amount of battery remaining.
+ */
 @property (readonly) uint batteryLevel;
+
+/*! An enum specifying the current selected palm rejection orientation (left vs. right)
+ * Deprecated in v2.0, moving forward please use palmRejectionOrientation
+*/
 @property (readwrite) JotPalmRejectionOrientation palmRejectionOrientation;
+
+/*! An enum specifying the current writing style and prefered writing hand. Default to JotWritingStyleRightDown
+ */
+@property (readwrite) JotWritingStyle writingStyle;
+
+/*! An enum specifying the current status of pairing styluses.
+ */
 @property (readonly) JotConnectionStatus connectionStatus;
+
+/*! An enum specifying the type of the preferred stylus.
+ * Deprecated in v2.0
+*/
 @property (readonly) JotPreferredStylusType preferredStylusType;
+
+/*! An enum specifying the preferred stylus.
+ */
+@property (readonly) JotPreferredStylus *preferredStylus;
+
+
 @end