+++ JotTouchSDK.framework/Versions/A/Headers/JotTouchMoment.h
@@ -11,10 +11,11 @@
 #import <UIKit/UIKit.h>
 @interface JotTouchMoment : NSObject
 @property (readwrite) CGPoint point;
-@property (readwrite) uint timestamp;
+@property (readwrite) NSTimeInterval timestamp;
 @property (readwrite) uint pressure;
 @property (readwrite) BOOL isTouchEnd;
 @property (readwrite) UITouch *touch;
--(id) initWithTouch:(UITouch *) touch withPoint:(CGPoint)point withPressure:(uint)pressure withTimestamp:(uint)timestamp;
+-(id) initWithTouch:(UITouch *) touch withPoint:(CGPoint)point withPressure:(uint)pressure withTimestamp:(NSTimeInterval)timestamp;
 +(JotTouchMoment *)touchEnd;
+
 @end